+++ spec/transformer_spec.rb
@@ -8,6 +8,11 @@ RSpec.describe Jekyll::WebAwesome::Transformer do
         This is info
         :::
 
+        >>>
+        ![Before](before.jpg)
+        ![After](after.jpg)
+        >>>
+
         ^^^
         Summary
         >>>
@@ -24,6 +29,9 @@ RSpec.describe Jekyll::WebAwesome::Transformer do
       result = described_class.process(markdown)
 
       expect(result).to include('<wa-callout variant="brand">')
+      expect(result).to include('<wa-comparison>')
+      expect(result).to include('<img slot="before" src="before.jpg" alt="Before" />')
+      expect(result).to include('<img slot="after" src="after.jpg" alt="After" />')
       expect(result).to include('<wa-details appearance=\'outlined\'>')
       expect(result).to include('<wa-tab-group placement="top">')
     end